Move delete button around
