+++ src/pages/edit-profile.tsx
@@ -3,14 +3,14 @@ import styles from "@/styles/Form.module.css";
 import { Form, Formik, FormikProps} from "formik";
 
 import useUser from "@/lib/useUser";
-import {Container} from "@mui/material"
+import {Button, Container} from "@mui/material"
 
 import * as Yup from 'yup';
 import { GetUserResponse } from "./api/userSession";
 import { KeyedMutator, useSWRConfig } from "swr";
 import { ScopedMutator } from "swr/_internal";
 import fetchJson from "@/lib/fetchJson";
-import Router from "next/router";
+import Router, { NextRouter, useRouter } from "next/router";
 import Head from "next/head";
 import { AcademicInterests, FirstName, GradeRange, LastName, Organization, Position, SciRENRegion, StatusAlert, SubmitButton, TextBio, UserType } from "@/components/FormComponents";
 
@@ -61,6 +61,7 @@ export const regionOptions = [
 
 const EditProfile = () => {
     const { mutate } = useSWRConfig()
+    const router = useRouter();
 
     const {user, mutateUser} = useUser((user) => "/login")
     return (
@@ -74,17 +75,23 @@ const EditProfile = () => {
             </Head>
             {user ? 
               user.isLoggedIn ?
-                editProfileForm(user, mutateUser, mutate):
+              (
+                <>
+                  {editProfileForm(user, mutateUser, mutate, router)}
+                </>
+              )
+                :
                 <h1>Loading</h1>
               :<h1>Loading</h1>  
             }
+            
       </div>
       </>
     )
   
 };
 
-const editProfileForm = (user: GetUserResponse, mutateUser: KeyedMutator<GetUserResponse>, mutate: ScopedMutator) =>{
+const editProfileForm = (user: GetUserResponse, mutateUser: KeyedMutator<GetUserResponse>, mutate: ScopedMutator, router: NextRouter) =>{
   const EditUserSchema = Yup.object().shape({
     userType: Yup.string()
         .required("Required"),
@@ -178,11 +185,35 @@ const editProfileForm = (user: GetUserResponse, mutateUser: KeyedMutator<GetUser
           <SciRENRegion regionOptions={regionOptions} touched={props.touched} errors={props.errors}/>
           <StatusAlert status={props.status}/>
           <SubmitButton/>
+          {deleteProfileButton(mutateUser, user, router)}
         </Container>
       </Form>
       )}
       
     </Formik>
+  )
+}
+
+
+const deleteProfileButton = (mutateUser: KeyedMutator<GetUserResponse>, user: GetUserResponse, router: NextRouter) => {
+  return (
+    <div className={styles.deleteUserContainer}>
+      <Button
+        variant="contained"
+        color="error"
+        className={styles.deleteUserButton}
+        onClick={async (e) => {
+          e.preventDefault();
+          fetch(`/api/user?userID=${user?.userID}`, { method: "DELETE" }), false;
+          console.log("test")
+          mutateUser(await fetchJson("/api/logout", { method: "POST" }), false);
+          router.replace("/login");
+        }}
+        type="submit"
+      >
+        Delete Account
+      </Button>
+    </div>
   )
 }
