+++ src/styles/Profile.module.css
@@ -53,9 +53,3 @@
 .profileHeading{
     margin-top: 10px
 }
-
-.deleteUser {
-    padding: 1rem;
-    width: 8%;
-    background: #e02121;
-  }
